Typo + move selenium test in its own file
As per Paul comment: https://github.com/gristlabs/grist-core/pull/651#discussion_r1318700964

"If you look in test/deployment you'll see that a selection of tests are imported there, including HomeIntro. These tests are used as a way to remotely check if a deployment of Grist is sane. In grist-core you can see these tests used in test/test_under_docker.sh, where a Grist server is run in a docker container, and then tests are run with the only access to the server being via the network.

Since your test relies on reconfiguring the server, it cannot be used as
a deployment test. So you'll need to move it to a different file."
